+++ packages/core/src/tracing/clsHooked/emitter-listener.js
@@ -4,6 +4,20 @@
  *
  * Original license: BSD 2 Clause
  * (see https://github.com/othiym23/emitter-listener/blob/7586fba839cf87774d5df224ce479c3b7e2b9964/package.json#L26)
+ *
+ * A note on the usage of the properties __wrappedInstana, __unwrapInstana, __wrapped  and __unwrap: The original
+ * package (https://github.com/othiym23/emitter-listener/blob/af156cef522f1daa11523b2d94b6f7fd1e859ca5/listener.js)
+ * attaches properties named __wrapped and __unwrap to the event emitter object. Even with this fixed and vendored
+ * version of emitter-listener, the issue fixed in https://github.com/othiym23/emitter-listener/pull/7 can still
+ * persist if other packages install the unfixed emitter-listener package as a dependency, and if the unfixed version
+ * of wrapEmitter gets called _after_ this version.
+ *
+ * To fix that, our version uses different property names (__wrappedInstana, __unwrapInstana) when attaching properties
+ * to the event emitter object. In addition to that, both the original and this fixed version use the package shimmer
+ * internally. The package shimmer also uses properties named __wrapped etc., but it attaches them to individual
+ * functions instead of the event emitter object. emitter-listener relies on this and checks whether a function has been
+ * shimmed by checking if the function has a __wrapped property. For those usages, we need to stick with the original
+ * property names.
  */
 // @ts-nocheck - not going to add type checking to a temporarily vendored dependency.
 
@@ -15,9 +29,7 @@ const shimmer = require('shimmer');
 const wrap = shimmer.wrap;
 const unwrap = shimmer.unwrap;
 
-// Default to complaining loudly when things don't go according to plan.
-// dunderscores are boring
-const SYMBOL = 'wrap@before';
+const SYMBOL = 'instana@before';
 
 // Sets a property on an object, preserving its enumerability.
 // This function assumes that the property is already writable.
@@ -161,18 +173,18 @@ module.exports = function wrapEmitter(emitter, onAddListener, onEmit) {
   // Only wrap `on` and `addListener` once. Correctly registering all
   // `onAddListener` callbacks for multiple consecutive wrapEmitter calls for
   // the same emitter is handled above by pushing them into an array if necessary.
-  if (!emitter.__wrapped) {
+  if (!emitter.__wrappedInstana) {
     wrap(emitter, 'addListener', adding);
     wrap(emitter, 'on', adding);
 
-    defineProperty(emitter, '__unwrap', function () {
+    defineProperty(emitter, '__unwrapInstana', function () {
       unwrap(emitter, 'addListener');
       unwrap(emitter, 'on');
       unwrap(emitter, 'emit');
       delete emitter[SYMBOL];
-      delete emitter.__wrapped;
+      delete emitter.__wrappedInstana;
     });
-    defineProperty(emitter, '__wrapped', true);
+    defineProperty(emitter, '__wrappedInstana', true);
   }
 
   // Always wrap `emit`, no matter if the emitter has already been wrapped by